Added more extensive logging, added blockChange and multiBlockChange
Still havent found source of error -3 but hope the new logging mechanism will help determine it in the future, as of block and multiblock changes, i have finally finished all (i think) of block update and loading mechanisms, so i will be finally start with gravity implementation

+++ file_io/file_io.h
@@ -13,7 +13,8 @@ char *file_io_read_buffer_from_file(const char *file_path, size_t *buffer_size);
 int file_io_write_formatted_to_file(const char *file_path, const char *format, ...);
 
 /*
+ * Appends
  * if buffer_size <= 0 returns without doing anything
  */
-int file_io_write_hex_to_file(const char *file_path, const char *buffer, size_t buffer_size);
+int file_io_write_hex_to_file(const char *file_path, const char *buffer, int buffer_size);
 #endif // FILE_IO_H
